An alternate implementation for supporting CombineFileInputFormat.
Now CombineFileInputFormat support is switched on via config in JobConf. MultiSourceTap support flows from it naturally.
