+++ cascading-hadoop/src/main/java/cascading/tap/hadoop/Hfs.java
@@ -26,12 +26,31 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Map;
 
+import org.apache.hadoop.fs.FileStatus;
+import org.apache.hadoop.fs.FileSystem;
+import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.fs.s3native.NativeS3FileSystem;
+import org.apache.hadoop.mapred.FileInputFormat;
+import org.apache.hadoop.mapred.FileOutputFormat;
+import org.apache.hadoop.mapred.InputSplit;
+import org.apache.hadoop.mapred.JobConf;
+import org.apache.hadoop.mapred.OutputCollector;
+import org.apache.hadoop.mapred.OutputLogFilter;
+import org.apache.hadoop.mapred.RecordReader;
+import org.apache.hadoop.mapred.Reporter;
+import org.apache.hadoop.mapred.lib.CombineFileInputFormat;
+import org.apache.hadoop.mapred.lib.CombineFileRecordReader;
+import org.apache.hadoop.mapred.lib.CombineFileSplit;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import cascading.flow.FlowProcess;
 import cascading.scheme.Scheme;
 import cascading.scheme.hadoop.SequenceFile;
 import cascading.tap.SinkMode;
 import cascading.tap.Tap;
 import cascading.tap.TapException;
+import cascading.tap.hadoop.io.CombineFileRecordReaderWrapper;
 import cascading.tap.hadoop.io.HadoopTupleEntrySchemeCollector;
 import cascading.tap.hadoop.io.HadoopTupleEntrySchemeIterator;
 import cascading.tap.type.FileType;
@@ -40,18 +59,6 @@ import cascading.tuple.TupleEntryCollector;
 import cascading.tuple.TupleEntryIterator;
 import cascading.tuple.hadoop.TupleSerialization;
 import cascading.util.Util;
-import org.apache.hadoop.fs.FileStatus;
-import org.apache.hadoop.fs.FileSystem;
-import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.fs.s3native.NativeS3FileSystem;
-import org.apache.hadoop.mapred.FileInputFormat;
-import org.apache.hadoop.mapred.FileOutputFormat;
-import org.apache.hadoop.mapred.JobConf;
-import org.apache.hadoop.mapred.OutputCollector;
-import org.apache.hadoop.mapred.OutputLogFilter;
-import org.apache.hadoop.mapred.RecordReader;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Class Hfs is the base class for all Hadoop file system access. Hfs may only be used with the
@@ -86,6 +93,9 @@ public class Hfs extends Tap<JobConf, RecordReader, OutputCollector> implements 
 
   /** Fields LOCAL_MODE_SCHEME * */
   public static final String LOCAL_MODE_SCHEME = "cascading.hadoop.localmode.scheme";
+
+  /** Field COMBINE_FILES_FOR_INPUT */
+  public static final String COMBINE_FILES_FOR_INPUT = "cascading.hadoop.hfs.combine.files";
 
   /** Field stringPath */
   protected String stringPath;
@@ -134,6 +144,41 @@ public class Hfs extends Tap<JobConf, RecordReader, OutputCollector> implements 
   protected static String getLocalModeScheme( JobConf conf, String defaultValue )
     {
     return conf.get( LOCAL_MODE_SCHEME, defaultValue );
+    }
+
+  /**
+   * Method setCombineFilesForInput provides a means to indicate whether to leverage
+   * {@link CombineFileInputFormat} for the input format. By default it is false.
+   *
+   * @param properties of type Map<Object,Object>
+   * @param combine    a boolean
+   */
+  public static void setCombineFilesForInput( Map<Object, Object> properties, boolean combine )
+    {
+    properties.put( COMBINE_FILES_FOR_INPUT, Boolean.toString( combine ) );
+    }
+
+  /**
+   * Method setCombineFilesForInput provides a means to indicate whether to leverage
+   * {@link CombineFileInputFormat} for the input format. By default it is false.
+   *
+   * @param conf    of type JobConf
+   * @param combine a boolean
+   */
+  public static void setCombineFilesForInput( JobConf conf, boolean combine )
+    {
+    conf.setBoolean( COMBINE_FILES_FOR_INPUT, combine );
+    }
+
+  /**
+   * Method getCombineFilesForInput returns whether to leverage {@link CombineFileInputFormat}
+   * for the input format. By default it is false.
+   *
+   * @param conf of type JobConf
+   */
+  private static boolean getCombineFilesForInput( JobConf conf )
+    {
+    return conf.getBoolean( COMBINE_FILES_FOR_INPUT, false );
     }
 
   protected Hfs()
@@ -370,6 +415,35 @@ public class Hfs extends Tap<JobConf, RecordReader, OutputCollector> implements 
     makeLocal( conf, qualifiedPath, "forcing job to local mode, via source: " );
 
     TupleSerialization.setSerializations( conf ); // allows Hfs to be used independent of Flow
+
+    // use CombineFileInputFormat if that is enabled
+    handleCombineFileInputFormat( conf );
+    }
+
+  /**
+   * Based on the configuration, handles and sets {@link CombineFileInputFormat} as the input
+   * format.
+   */
+  private void handleCombineFileInputFormat( JobConf conf )
+    {
+    // if combining files, override the configuration to use CombineFileInputFormat
+    if( getCombineFilesForInput( conf ) )
+      {
+      // get the prescribed individual input format from the underlying scheme so it can be used by CombinedInputFormat
+      String individualInputFormat = conf.get( "mapred.input.format.class" );
+      if( individualInputFormat == null )
+        throw new TapException( "input format is missing from the underlying scheme" );
+
+      if( individualInputFormat.equals( CombinedInputFormat.class.getName() ) &&
+          conf.get( CombineFileRecordReaderWrapper.INDIVIDUAL_INPUT_FORMAT ) == null )
+        throw new TapException( "the input format class is already the combined input format but the underlying input format is missing" );
+
+      // set the underlying individual input format
+      conf.set( CombineFileRecordReaderWrapper.INDIVIDUAL_INPUT_FORMAT, individualInputFormat );
+
+      // override the input format class
+      conf.setInputFormat( CombinedInputFormat.class );
+      }
     }
 
   @Override
@@ -597,5 +671,14 @@ public class Hfs extends Tap<JobConf, RecordReader, OutputCollector> implements 
       return;
 
     statuses = getFileSystem( conf ).listStatus( getPath() );
+    }
+
+  /** Combined input format that uses the underlying individual input format to combine multiple files into a single split. */
+  static class CombinedInputFormat extends CombineFileInputFormat
+    {
+    public RecordReader getRecordReader( InputSplit split, JobConf job, Reporter reporter ) throws IOException
+      {
+      return new CombineFileRecordReader( job, (CombineFileSplit) split, reporter, CombineFileRecordReaderWrapper.class );
+      }
     }
   }
